+++ .gitignore
@@ -57,4 +57,7 @@ $RECYCLE.BIN/
 # Windows shortcuts
 *.lnk
 
+# webpack
+node_modules/
+
 # End of https://www.gitignore.io/api/osx,windows
