+++ diplom/index.html
@@ -19,7 +19,7 @@
 
     <title>Септик-погреб</title>
 
-    <script src="./js/script.js" defer></script>
+    <script src="./dist/main.js" defer></script>
 </head>
 
 <body>
