+++ diplom/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "glodiplom",
+  "version": "0.0.1",
+  "description": "JS11 Дипломный проект",
+  "main": "index.js",
+  "scripts": {
+    "dev": "webpack-dev-server --open --mode=development",
+    "build": "webpack --mode=production"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Ckopn84/GloAcademy/tree/diplom/diplom"
+  },
+  "author": "Бускин И.Н.",
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/core": "^7.8.7",
+    "@babel/preset-env": "^7.8.7",
+    "babel-loader": "^8.0.6",
+    "path": "^0.12.7",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
+  }
+}
