+++ diplom/src/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+import togglePopUp from './modules/togglePopUp';
+import showSentenceBlock from './modules/showSentenceBlock';
+
+togglePopUp();
+showSentenceBlock();
