+++ diplom/src/modules/showSentenceBlock.js
@@ -1,0 +1,18 @@
+
+const showSentenceBlock = () => {
+	const sentenceRow = document.querySelector('.sentence .row'),
+		addSentenceBtn = document.querySelector('.add-sentence-btn');
+	const blocks = document.querySelectorAll('.shadow-block');
+
+	addSentenceBtn.addEventListener('click', () => {
+		for (const elem of sentenceRow.children) {
+			if (elem.classList.contains('visible-sm-block'))
+				elem.classList.remove('visible-sm-block');
+			if (elem.classList.contains('hidden'))
+				elem.classList.remove('hidden');
+			addSentenceBtn.style.display = 'none';
+		}
+	});
+};
+
+export default showSentenceBlock;
