+++ diplom/webpack.config.js
@@ -1,0 +1,25 @@
+
+const path = require('path');
+
+module.exports = {
+	entry: {
+		main: './src/index.js'
+	},
+	output: {
+		path: path.resolve(__dirname, './dist'),
+		filename: '[name].js',
+		publicPath: '/dist'
+	},
+	devServer: {
+		overlay: true,
+	},
+	module: {
+		rules: [
+			{
+				test: /\.js$/,
+				loader: 'babel-loader',
+				exclude: '/node_modules/'
+			}
+		]
+	}
+};
